create AbstractUser
